+++ examples/example.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta content="width=device-width,height=device-height,initial-scale=1" name="viewport">
+  </head>
+  <body>
+    <canvas id="test"></canvas>
+
+    <script src="../dist/index.js"></script>
+
+    <script type="application/javascript">
+      const manager = new window.CanvasManager(document.querySelector('#test'));
+    </script>
+  </body>
+</html>
